requirements.txt
# DeepAgents
deepagents>=0.1.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database Connectors
sqlalchemy>=2.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0
clickhouse-driver>=0.2.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Caching
redis>=5.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
